change data type for lesson number field

+++ calendar_plan/src/main/java/com/ejournal/calendar_plan/dto/CalendarPlanRecordRequestDto.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 public class CalendarPlanRecordRequestDto {
 
-    private Long lessonNumber;
+    private Integer lessonNumber;
     private String themeName;
     private String individualAssignment;
     private Date individualAssignmentDate;
